Update the OOB experience by having a 'welcome! change your password' page.

+++ core/user/helpers/user_helpers.rb
@@ -1,19 +1,24 @@
 class Main
   helpers do
+    # Overridden in aura to track last login time.
+    # This is called when a user logs in.
+    #
     def redirect_to_return_url(session_key = :return_to, default = '/admin')
-      redirect session.delete(:return_to) || default
-    end
+      u = current_user
+      puts "-"*80
+      p u.inspect
+      first_login = u.last_login.nil?
 
-    def authenticate(opts)
-      if user = __USER__.authenticate(opts[:username], opts[:password])
-        return  if session[:user] == user.id
-        session[:user] = user.id
+      u.last_login = Time.now
+      u.save
 
-        # Added in Aura.
-        u = current_user
-        u.last_login = Time.now
-        u.save
+      # On a user's first login, go to a admin welcome page
+      # where they can change their password.
+      if first_login
+        redirect R(:admin, :welcome)
       end
+
+      redirect session.delete(:return_to) || default
     end
   end
 end
